requirements.txt
anyio==4.4.0
awsebcli==3.20.10
bcrypt==4.2.0
botocore==1.31.85
cement==2.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
fastapi==0.100.0
h11==0.14.0
idna==3.7
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.5
motor==3.1.1
passlib==1.7.4
pathspec==0.10.1
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.8
pymongo==4.8.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pywin32==306
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
semantic-version==2.8.5
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
termcolor==1.1.0
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.23.2
wcwidth==0.1.9
